src/requirements.txt
asgiref==3.4.1
backend==0.2.4.1
certifi==2021.5.30
cffi==1.14.6
charset-normalizer==2.0.4
crispy==0.6.3
cryptography==3.4.8
cycler==0.10.0
defusedxml==0.7.1
Django==3.2.6
django-allauth==0.45.0
django-crispy-forms==1.12.0
django-filter==2.4.0
django-widget-tweaks==1.4.8
djangorestframework==3.12.4
fabio==0.12.0
h5py==3.4.0
idna==3.2
kiwisolver==1.3.2
matplotlib==3.4.3
numpy==1.21.2
oauthlib==3.1.1
Pillow==8.3.1
pycparser==2.20
PyJWT==2.1.0
pyparsing==2.4.7
PyQt5==5.15.4
PyQt5-Qt5==5.15.2
PyQt5-sip==12.9.0
python-dateutil==2.8.2
python3-openid==3.2.0
pytz==2021.1
requests==2.26.0
requests-oauthlib==1.3.0
silx==0.15.2
six==1.16.0
sqlparse==0.4.1
urllib3==1.26.6
